random mvie button fix
Issues #346 #25

Random movie button had an issue because "format()" function inside "butter-movies-provider/index.js" was expecting to get an array but "GET /random/movie" is sending an object so there is an issue because ".foreach()" inside that "format()" function is working only with arrays.

Now it is possible to fix "butter-movies-provider/index.js" which could be just a couple of code lines or to change "filter_bar.js" which require to copy and change 3 functions from "butter-movies-provider/index.js".

Above is my proposal for changes for "filter_bar.js". There are 3 new functions that I have been copied and changed from "butter-movies-provider/index.js" into "filter_bar.js".